microservice/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]==0.115.12",
    "nh3==0.2.21",
    "passlib[bcrypt]==1.7.4",
    "pydantic-settings==2.9.1",
    "pymongo==4.13.0",
    "python-dotenv==1.1.0",
    "python-jose==3.4.0",
]
